+++ capston2-tetris/1Pindex.html
@@ -21,6 +21,6 @@
         </div>
         <div id="divNextImg1"></div>
     </div>
-    <script src="/1PtetrisUpdate.js" type="module"></script>
+    <script src="/tetrisUpdate.js" type="module"></script>
 </body>
 </html>
