+++ capston2-tetris/2Pindex.html
@@ -16,16 +16,16 @@
             <div class="player1">
             <div class="score">0</div>
             <div class="playground">
+                <div id="divNextImg1"></div>
                 <ul></ul>
             </div>
-            <div id="divNextImg1"></div>
             </div>
             <div class="player2">
             <div class="Secondscore">0</div>
             <div class="Secondplayground">
+                <div id="divNextImg2"></div>
                 <ul></ul>
             </div>
-            <div id="divNextImg2"></div>
             </div>
     </div>
     <script src="/1PtetrisUpdate.js" type="module"></script>
