+++ capston2-tetris/public/css/1Pstyle.css
@@ -2,6 +2,11 @@
     margin: 0;
     padding: 0;
 }
+
+.wrapper {
+    text-align: center;
+}
+
 ul{
     list-style: none;
 }
@@ -31,17 +36,17 @@ body {
     background: salmon;
     border: none;
     cursor:pointer;
-    margin-top: 10px;
-    margin-bottom: 15px;
 }
 .score{
     text-align: center;
     font-size: 36px;
-    margin-bottom: 2rem;
+    margin-bottom: 10px;
+    margin-top: 3rem;
 }
 
 /* playground의 ul에 줌 */
 .playground{
+    display: inline-block;
     border: 1px;
 }
 .playground > ul{
@@ -62,6 +67,12 @@ body {
     width: 25px;
     height: 25px;
     border: 1px solid #fff;
+}
+
+#divNextImg1 {
+    position: absolute;
+    padding-left: 20px;
+    display: inline;
 }
 
 .tree{
