+++ capston2-tetris/public/js/tetrisUpdate.js
@@ -3,8 +3,7 @@ import BLOCKS from "./blocks.js";
 const playground = document.querySelector(".playground > ul");
 const gameText = document.querySelector(".game-Text");
 const scoreDisplay = document.querySelector(".score");
-const restartButton = document.querySelector(".game-Text .restart");
-
+const restartButton = document.querySelector(".game-Text > button");
 
 // Setting
 const GAME_ROWS = 20;
@@ -13,8 +12,9 @@ const GAME_COLS = 10;
 
 // variables
 let score = 0; // 점수
-let duration = 800; // 내려가는 속도
+let duration = 1000; // 내려가는 속도
 let downInterval; // 
+let levelstack = 0;
 let tempMovingItem; // 무빙을 실질적으로 실행하기 전에 잠깐 담아두는 용도
 let nextMovingItem = null;
 
@@ -29,7 +29,7 @@ init()
 
 // functions
 function init(){
-    //tempMovingItem = {...movingItem}; // ... 사용하면 값만 가져와서 넣음 나중에 변경이 되어도 tempMovingItem은 변하지 않음
+    tempMovingItem = {...movingItem}; // ... 사용하면 값만 가져와서 넣음 나중에 변경이 되어도 tempMovingItem은 변하지 않음
     for(let i = 0; i < GAME_ROWS; i++){ // 가로 10칸짜리 블록 세로 방향으로 20번 만들기
         prependNewLine()
     }
@@ -68,9 +68,13 @@ function renderBlocks(moveType=""){
         // const xxx = 조건 ? 참일 경우 : 거짓일 경우
         const target = playground.childNodes[y] ? playground.childNodes[y].childNodes[0].childNodes[x] : null;
         const isAvailable = checkEmpty(target);
+        
         if(isAvailable){
+            // console.log(tempMovingItem)
+            // console.log(moveType)
             target.classList.add(type, "moving")
         }
+ 
         else{
             tempMovingItem = {... movingItem}
             if(moveType === 'retry'){
@@ -86,9 +90,12 @@ function renderBlocks(moveType=""){
             return true;
         }
     }); // tree의 0번째 인덱스 다 갖고옴
+    movingItem.type = type;
     movingItem.left = left;
     movingItem.top = top;
     movingItem.direction = direction;
+
+    speedUp();
 }
 
 function seizeBlock(){
@@ -98,7 +105,6 @@ function seizeBlock(){
         moving.classList.add("seized");
     })
     checkMatch()
-    
 }
 
 function checkMatch(){
@@ -114,11 +120,10 @@ function checkMatch(){
             child.remove();
             prependNewLine();
             score++;
+            levelstack++;
             scoreDisplay.innerText = score;
         }
     })
-    tempMovingItem = nextMovingItem
-    nextMovingItem = null
     generateNewBlock()
 }
 function generateNewBlock(){
@@ -126,16 +131,21 @@ function generateNewBlock(){
     downInterval = setInterval(()=>{
         moveBlock('top', 1)
     },duration)
+    tempMovingItem = nextMovingItem
+    nextMovingItem = null
 
     const blockArray = Object.entries(BLOCKS);
     const randomInex = Math.floor(Math.random() * blockArray.length);
-    movingItem.type =blockArray[randomInex][0];
-    movingItem.top = 0;
-    movingItem.left = 3;
-    movingItem.direction = 0;
-    tempMovingItem = {...movingItem}
-    console.log(tempMovingItem)
-    console.log(nextMovingItem)
+    if(tempMovingItem == null){
+        movingItem.type =blockArray[randomInex][0];
+        movingItem.top = 0;
+        movingItem.left = 3;
+        movingItem.direction = 0;
+        tempMovingItem = {...movingItem}
+        console.log(duration);
+        console.log(levelstack);
+        speedUp();
+    }
     if(nextMovingItem == null){
         makeNewBlock()
     }
@@ -150,7 +160,7 @@ function makeNewBlock(){
     movingItem.left = 3;
     movingItem.direction = 0;
     nextMovingItem = {...movingItem}
-    console.log(nextMovingItem)
+    
     showNextBlock()
 }
 function checkEmpty(target){
@@ -175,27 +185,35 @@ function dropBlock(){
     clearInterval(downInterval);
     downInterval = setInterval(()=>{
         moveBlock("top", 1)
-    }, 100)
+    }, 4)
+}
+
+function speedUp() {
+    if(levelstack > 11) {
+        duration = duration - 160;
+        levelstack = 0;
+    }
 }
 
 function showGameOverText(){
     gameText.style.display = "flex"
+    divNextImg1.style.zIndex = "-1"
 }
 // event handling
 
 
 document.addEventListener("keydown", e=>{
     switch(e.keyCode){
-        case 68:
+        case 39:
             moveBlock("left", 1);
             break;
-        case 65:
+        case 37:
             moveBlock("left", -1);
             break;
-        case 83:
+        case 40:
             moveBlock("top", 1);
             break;    
-        case 87:
+        case 38:
             changeDirection();
             break;
         case 32:
@@ -211,22 +229,7 @@ restartButton.addEventListener("click",()=>{
     scoreDisplay.innerText = 0; // 다시 시작 버튼 눌렀을 때 스코어 0으로 초기화
     init()
 })
-ranking.addEventListener('click',()=>{
-    let name = prompt('이름을 입력해주세요')
-})
 
 function showNextBlock() {
-    $("#divNextImg").html(`<img src = "/${nextMovingItem.type}.png">`);
+    $("#divNextImg1").html(`<img src = "/${nextMovingItem.type}.png">`);
 }
-
-
-
-var sound = new Howl({
-    src: ['/plum.mp3'],
-    volume: 0.5,
-    autoplay: true,
-    loop: true,
-    onend : () => {
-    console.log('Finished!');
-    }
-});
