+++ capston2-tetris/views/Lobby.ejs
@@ -39,20 +39,17 @@
                     ↑ : 블럭 회전<br>
                     ↓ : 소프트 드롭<br>
                     space : 하드 드롭<br>
-                    ????? : 아이템 사용<br>
                     <div class="line"></div>
                     <p class="pophead">2인 멀티플레이 키</p>
                     1p A D : 블럭 좌우 이동<br>
                     1p W : 블럭 회전<br>
                     1p S : 소프트 드롭<br>
-                    1p ????? : 하드 드롭<br>
-                    1p ????? : 아이템 사용<br>
+                    1p space : 하드 드롭<br>
                     <br>
                     2p ← → : 블럭 좌우 이동<br>
                     2p ↑ : 블럭 회전<br>
                     2p ↓ : 소프트 드롭<br>
-                    2p ????? : 하드 드롭<br>
-                    2p ????? : 아이템 사용<br>
+                    2p num0, / : 하드 드롭<br>
                     <div class="line"></div>
                 </div>
             </div>
